+++ src/main/java/com/syspa/login_service/repository/RefreshTokenRepository.java
@@ -1,0 +1,13 @@
+package com.syspa.login_service.repository;
+
+import com.syspa.login_service.model.RefreshToken;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+  Optional<RefreshToken> findByToken(String token);
+
+  void deleteByUsername(String username);
+}